+++ src/SoundDeck.PI/js/components/samplerSettings.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Checkbox, FolderPicker, PropertyInspectorWrapper, Select } from "react-sharpdeck";
+import { Checkbox, FolderPicker, PropertyInspectorWrapper, Range, Select } from "react-sharpdeck";
 import PlayActionTypes from "../models/playActionTypes";
 
 class SamplerSettings extends React.Component {
@@ -24,6 +24,7 @@ class SamplerSettings extends React.Component {
                 <hr />
                 <Select label="Playback Device" dataSourceUri="GetPlaybackAudioDevices" valuePath="playbackAudioDeviceId" defaultValue="PLAYBACK_DEFAULT" />
                 <Select label="Action" options={this.actionTypes} valuePath="action" defaultValue={PlayActionTypes.PlayNext.value} />
+                <Range label="Volume" valuePath="playbackVolume" defaultValue="75" min="0" max="100" step="1" />
             </PropertyInspectorWrapper>
         );
     }